Upgrade the prefix_scan() function to operate in log2(n) time
By updating the algorithm we increase the speedup at the expense of
requiring a memcpy per iteration as we are no longer adding the
original values to our output sums.

If we didn't have the memcpy() to copy values to read from, then the
parallelisation would be unsafe as the updates would depend on other
values of the array which may already be updated by another thread.